resolve circular dependency issue
