+++ tensorboard/webapp/metrics/actions/BUILD
@@ -10,7 +10,7 @@ tf_ts_library(
         "index.ts",
     ],
     deps = [
-        "//tensorboard/webapp/metrics:internal_types",
+        "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/data_source",
         "//tensorboard/webapp/metrics/views/card_renderer:scalar_card_types",
         "//tensorboard/webapp/util:dom",
