+++ tensorboard/webapp/metrics/store/BUILD
@@ -18,7 +18,7 @@ tf_ts_library(
         "//tensorboard/webapp/app_routing:types",
         "//tensorboard/webapp/app_routing/actions",
         "//tensorboard/webapp/core/actions",
-        "//tensorboard/webapp/metrics:internal_types",
+        "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics:utils",
         "//tensorboard/webapp/metrics/actions",
         "//tensorboard/webapp/metrics/data_source",
@@ -68,7 +68,7 @@ tf_ts_library(
     ],
     deps = [
         "//tensorboard/webapp/app_routing:namespaced_state_reducer_helper",
-        "//tensorboard/webapp/metrics:internal_types",
+        "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/data_source",
         "//tensorboard/webapp/metrics/views/card_renderer:scalar_card_types",
         "//tensorboard/webapp/types",
@@ -92,8 +92,8 @@ tf_ts_library(
         "//tensorboard/webapp/app_routing:types",
         "//tensorboard/webapp/app_routing/actions",
         "//tensorboard/webapp/core/actions",
-        "//tensorboard/webapp/metrics:internal_types",
         "//tensorboard/webapp/metrics:test_lib",
+        "//tensorboard/webapp/metrics:types",
         "//tensorboard/webapp/metrics/actions",
         "//tensorboard/webapp/metrics/data_source",
         "//tensorboard/webapp/persistent_settings",
