+++ tensorboard/webapp/metrics/store/metrics_selectors_test.ts
@@ -15,7 +15,6 @@ limitations under the License.
 import {DataLoadState} from '../../types/data';
 import {nextElementId} from '../../util/dom';
 import {PluginType} from '../data_source';
-import {HistogramMode, TooltipSort, XAxisType} from '../internal_types';
 import {
   appStateFromMetricsState,
   buildMetricsSettingsState,
@@ -26,6 +25,7 @@ import {
   createScalarStepData,
   createTimeSeriesData,
 } from '../testing';
+import {HistogramMode, TooltipSort, XAxisType} from '../types';
 import * as selectors from './metrics_selectors';
 
 describe('metrics selectors', () => {
