+++ tensorboard/webapp/metrics/views/card_renderer/scalar_card_component.ng.html
@@ -137,7 +137,7 @@ limitations under the License.
       <tbody>
         <ng-container
           *ngFor="
-            let datum of getCursorAwareTooltipData(tooltipData, 
+            let datum of getCursorAwareTooltipData(tooltipData,
               cursorLocationInDataCoord, cursorLocation);
             trackBy: trackByTooltipDatum
           "
