+++ tensorboard/webapp/persistent_settings/_data_source/BUILD
@@ -10,6 +10,7 @@ tf_ng_module(
     ],
     deps = [
         ":types",
+        "//tensorboard/webapp/metrics:types",
         "@npm//@angular/core",
         "@npm//@ngrx/store",
         "@npm//rxjs",
@@ -20,6 +21,9 @@ tf_ng_module(
     name = "types",
     srcs = [
         "types.ts",
+    ],
+    deps = [
+        "//tensorboard/webapp/metrics:types",
     ],
 )
 
@@ -48,6 +52,7 @@ tf_ng_module(
         ":_data_source",
         ":types",
         "//tensorboard/webapp/angular:expect_angular_core_testing",
+        "//tensorboard/webapp/metrics:types",
         "@npm//@angular/core",
         "@npm//@types/jasmine",
         "@npm//rxjs",
