+++ tensorboard/webapp/persistent_settings/_data_source/types.ts
@@ -13,23 +13,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
+import {TooltipSort} from '../../metrics/types';
+
 export enum ThemeValue {
   BROWSER_DEFAULT = 'browser_default',
   LIGHT = 'light',
   DARK = 'dark',
-}
-
-// When adding a new value to the enum, please implement the deserializer on
-// data_source/metrics_data_source.ts.
-// When editing a value of the enum, please write a backward compatible
-// deserializer in tensorboard/webapp/metrics/store/metrics_reducers.ts
-export enum TooltipSort {
-  DEFAULT = 'default',
-  ALPHABETICAL = 'alphabetical',
-  ASCENDING = 'ascending',
-  DESCENDING = 'descending',
-  NEAREST = 'nearest',
-  NEAREST_Y = 'nearest_Y',
 }
 
 /**
